+++ docs/source/conf.py
@@ -27,9 +27,9 @@ copyright = '2018-2023, Berlin Institute of Health'
 author = 'BIH Core Unit Bioinformatics'
 
 # The short X.Y version
-version = '0.12'
+version = '0.13'
 # The full version, including alpha/beta/rc tags
-release = '0.12.0'
+release = '0.13.0-WIP'
 
 
 # -- General configuration ---------------------------------------------------